The Fundamentals of C: Challenge: Build the foundation
This is much more than required by the challenge. I am inferring the requirements of the final challenge. It seems related to memory registers and I'm not sure if there is a better way of referencing them compared to iterating through letters in the alphabet.

The only validation I perform is if the input is an integer. No feedback is provided if the integer is out of range or if the input is an invalid type.

By including an if/else statement within a for-loop, that evaluation is performed with each loop. This is unnecessary but can be separated out later into its own function.